Updates for portability. Updated error reporting. Added OS arguments.

+++ me/game/Game.h
@@ -102,7 +102,7 @@ namespace me::game
 		int GetComponentCount() const override;
 		void AddComponent( game::component::IGameComponent::ptr component ) override;
 		void RemoveComponent( game::component::IGameComponent::ptr component ) override;
-		game::component::IGameComponent::ptr GetComponent( int index ) override;
+		game::component::IGameComponent::ptr GetComponent( size_t index ) override;
 		game::component::IGameComponent::ptr GetComponent( std::string typeName ) override;
 		int FindComponent( std::string typeName ) const override;
