+++ me/scene/Scene.cpp
@@ -198,11 +198,11 @@ void Scene::RemoveComponent( component::ISceneComponent::ptr component )
 	component->OnDetach( this );
 }
 
-component::ISceneComponent* Scene::GetComponent( int index )
+component::ISceneComponent* Scene::GetComponent( size_t index )
 {
-	if ( index > (int)m_components.size() ) return nullptr;
+	if ( index > m_components.size() ) return nullptr;
 
-	int i = 0;
+	size_t i = 0;
 	for ( auto component : m_components )
 	{
 		if ( index == i ) return component.get();
